Make SXT-solo config use base NN as IP (stop adding 100) (#49)
* Make SXT-core use base NN as IP (stop adding 100)

This removes adding 100 to the final octet of the assigned IP address when using a SXT as the core router. Adding 100 should be reserved for a second router at a node, whereas, in this config, the SXT is meant to act as the first (core) router.

* Make SXT-solo use base NN as IP (stop adding 100)

This removes adding 100 to the final octet of the assigned IP address when using the SXT-solo config. We want to make sure that the NN-based IP addresses of an active node will always resolve.